Added DriverTests and refactored to get compatibility
