--- csharp/test/Drivers/Apache/Impala/ImpalaTests.cs
@@ -1,44 +1,0 @@
-﻿/*
-* Licensed to the Apache Software Foundation (ASF) under one or more
-* contributor license agreements.  See the NOTICE file distributed with
-* this work for additional information regarding copyright ownership.
-* The ASF licenses this file to You under the Apache License, Version 2.0
-* (the "License"); you may not use this file except in compliance with
-* the License.  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-
-using System.Collections.Generic;
-using Apache.Arrow.Adbc.Drivers.Apache.Impala;
-using Apache.Arrow.Adbc.Tests.Xunit;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Impala
-{
-    [TestCaseOrderer("Apache.Arrow.Adbc.Tests.Xunit.TestOrderer", "Apache.Arrow.Adbc.Tests")]
-    public class ImpalaTests : TestBase<ApacheTestConfiguration, ImpalaTestEnvironment>
-    {
-        public ImpalaTests(ITestOutputHelper? outputHelper)
-            : base(outputHelper, new ImpalaTestEnvironment.Factory())
-        {
-        }
-
-        [SkippableFact, Order(1)]
-        public void CanExecuteQuery()
-        {
-            AdbcStatement statement = Connection.CreateStatement();
-            statement.SqlQuery = TestConfiguration.Query;
-            QueryResult queryResult = statement.ExecuteQuery();
-
-            DriverTests.CanExecuteQuery(queryResult, TestConfiguration.ExpectedResultsCount);
-        }
-    }
-}
